fixed the UI for the percentage of song complete.

+++ Assets/Scripts/Save Data/PersistentDataInformation.cs
@@ -32,7 +32,7 @@ public class PersistentDataInformation
         public int m_stars;
         public int m_totalNote;
         public int m_notesHit;
-        public int m_songCompletionPercentage;
+        public float m_songCompletionPercentage;
         public string _Difficulty;
     }
